auto

+++ src/pci.c
@@ -1,7 +1,7 @@
 #include <system.h>
 
 void detectDevices(){
-	unsigned int i = 0;
+	unsigned long i = 0;
 	for(unsigned char busses = 0 ; busses < 250 ; busses++){
 		for(unsigned char slots = 0 ; slots < 32 ; slots++){
 			for(unsigned char functions = 0 ; functions < 8 ; functions++){
